Fixes a bad example.
The example with the socks which tried to resemble an exponential
distribution is incorrect. It specified the *quantity of socks* needed in
order for a microorganism empire to flourish. The exponential
distribution is concerned with continuous random variables. Therefore
measuring the *area* of the room covered with dirty socks is more
appropriate.